+++ app/datatables/index_datatable.rb
@@ -3,11 +3,12 @@ class IndexDatatable < AjaxDatatables::Datatable
 
   class Filters < AjaxDatatables::Filters; end
   def manipulate(records)
-    super(records)
-    super(records).where(build_conditions(filter_searching_nodes))
+    super(records).where(build_conditions(filter_searching_nodes)).order(ordering_sql)
   end
 
   def filter_searching_nodes
+    return [] if view_context.nil?
+
     columns.select(&:searchable).map do |column|
       sv = params[column.name].presence || next
       { column_name: column.name, search_value: sv }
@@ -18,6 +19,16 @@ class IndexDatatable < AjaxDatatables::Datatable
     @filters ||= "#{default_model.to_s.pluralize}Datatable::Filters".constantize.new(datatable: self)
   rescue NameError
     nil
+  end
+
+  def ordering_sql
+    return nil if view_context.nil?
+    if (column = columns[params[:sort_column]])
+      direction = (params[:sort_direction] == 'desc') ? :desc : :asc
+      ["#{column.source} #{direction}"]
+    else
+      nil
+    end
   end
 
   def default_settings