+++ spec/rails_helper.rb
@@ -9,6 +9,7 @@ ActiveRecord::Migration.maintain_test_schema!
 
 require 'spec_helper'
 require 'ajax_feature_helper'
+require 'controller_spec_helper'
 
 Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
 
@@ -48,4 +49,5 @@ RSpec.configure do |config|
   config.include AjaxFeatureHelper
   config.include AjaxFeatureHelper::Filter
   config.include AjaxFeatureHelper::Order
+  config.include ControllerSpecHelper
 end