+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - accept_categories: parser or updater?
 - json:: depends to serverside ??
 - limit: spec
+- spec: owg name correction
+- sorting_sql: super
  
 ### new feature